Pull Req: Refactored legacy code folder structure
