+++ src/components/NavUI/Nav.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import Container from 'react-bootstrap/Container';
 import Nav from 'react-bootstrap/Nav';
 import Navbar from 'react-bootstrap/Navbar';
-import logo from '../assets//logos/logo.png';
+import logo from '../../assets/logos/logo.png';
 import { NavLink } from 'react-router-dom';
