+++ src/components/SummaryUI/Summary.css
@@ -25,7 +25,7 @@
     left: 0;
     width: 100%;
     height: 100%;
-    background-image: url('../assets/backgrounds/lights.jpg');
+    background-image: url('../../assets/backgrounds/lights.jpg');
     background-size: cover;
     background-position: center;
     background-repeat: no-repeat;
